+++ examples/hosted_dashboard/.terraform.lock.hcl
@@ -1,0 +1,10 @@
+# This file is maintained automatically by "terraform init".
+# Manual edits may be lost in future updates.
+
+provider "locally/debug/coralogix" {
+  version     = "1.3.0"
+  constraints = "~> 1.3"
+  hashes = [
+    "h1:Jth2b98HR3OiB1MpnQFenI13TVjzCyNT+rsBUr1L7bA=",
+  ]
+}
