+++ examples/hosted_dashboard/grafana_acc_dashboard.json
