+++ examples/hosted_dashboard/grafana_acc_updated_dashboard.json
