+++ examples/hosted_dashboard/grafana_dashboard.json
